add signup to bounty popup
